Add clarification regarding test data changesets.
As a new user i found the test data section a little bit confusing as it led me to believe that simply by not specifying a `"test"` context then changesets marked with `"test"` context would not be run. As i understand, this is only true if another context is being specified, but when no context is specified all changesets are applied, including those marked with `"test"` context. I have added a clarification about this.